+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
   * Upgraded chai@4.0.2, chai-as-promised@7.0.0, promise-duplex@0.1.2,
     snazzy@7.0.0, standard@10.0.2, tap@10.5.1, tap-given@0.4.1
+  * Bugfixes for tests
 
 ## v0.0.1 2017-03-12